+++ src/lib.rs
@@ -62,14 +62,6 @@
     doc = "`tokio-02` (*inactive*) | `tokio::io::AsyncBufRead`, `tokio::io::AsyncWrite`"
 )]
 #![cfg_attr(
-    feature = "tokio-03",
-    doc = "[`tokio-03`](crate::tokio_03) | [`tokio::io::AsyncBufRead`](::tokio_03::io::AsyncBufRead), [`tokio::io::AsyncWrite`](::tokio_03::io::AsyncWrite)"
-)]
-#![cfg_attr(
-    not(feature = "tokio-03"),
-    doc = "`tokio-03` (*inactive*) | `tokio::io::AsyncBufRead`, `tokio::io::AsyncWrite`"
-)]
-#![cfg_attr(
     feature = "tokio",
     doc = "[`tokio`](crate::tokio) | [`tokio::io::AsyncBufRead`](::tokio::io::AsyncBufRead), [`tokio::io::AsyncWrite`](::tokio::io::AsyncWrite)"
 )]
@@ -178,9 +170,6 @@ pub mod tokio;
 #[cfg(feature = "tokio-02")]
 #[cfg_attr(docsrs, doc(cfg(feature = "tokio-02")))]
 pub mod tokio_02;
-#[cfg(feature = "tokio-03")]
-#[cfg_attr(docsrs, doc(cfg(feature = "tokio-03")))]
-pub mod tokio_03;
 
 mod unshared;
 mod util;