+++ tests/utils/algos.rs
@@ -81,9 +81,6 @@ macro_rules! algos {
                 #[cfg(feature = "tokio-02")]
                 io_algo!(tokio_02, $name($encoder, $decoder));
 
-                #[cfg(feature = "tokio-03")]
-                io_algo!(tokio_03, $name($encoder, $decoder));
-
                 #[cfg(feature = "tokio")]
                 io_algo!(tokio, $name($encoder, $decoder));
             }