+++ tests/utils/mod.rs
@@ -3,8 +3,6 @@
 mod input_stream;
 #[cfg(feature = "tokio-02")]
 mod tokio_02_ext;
-#[cfg(feature = "tokio-03")]
-mod tokio_03_ext;
 #[cfg(feature = "tokio")]
 mod tokio_ext;
 mod track_closed;