+++ tests/utils/test_cases.rs
@@ -697,9 +697,6 @@ macro_rules! test_cases {
             #[cfg(feature = "tokio-02")]
             io_test_cases!(tokio_02, $variant);
 
-            #[cfg(feature = "tokio-03")]
-            io_test_cases!(tokio_03, $variant);
-
             #[cfg(feature = "tokio")]
             io_test_cases!(tokio, $variant);
         }